API/bigquery/vids/Python3/GIS/requirements.txt
google-cloud-bigquery 
google-cloud-bigquery-datatransfer 
tornado
watchdog 
shapely
geojson